+++ index.js
@@ -1,7 +1,6 @@
 import express from "express"
 import graphqlHTTP from "express-graphql"
-import schema from "./schema"
-import resolver from "./resolvers"
+import { schema } from "./data/db/schema"
 
 const app = express()
 
@@ -9,11 +8,11 @@ app.get('/', (req, res) => {
     res.send("GraphQL learning in progress")
 })
 
-const root = resolver;
+
 
 app.use('/graphql', graphqlHTTP({
     schema: schema,
-    rootValue: root,
+
     graphiql: true
 }))
