+++ package.json
@@ -25,6 +25,7 @@
     "express-graphql": "^0.9.0",
     "graphql": "^15.0.0",
     "graphql-tools": "^6.0.8",
+    "mongoose": "^5.9.18",
     "nodemon": "^2.0.4"
   }
 }
